disambiguate a logger call
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@254 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ pdns/distributor.hh
@@ -198,7 +198,7 @@ template<class Answer, class Question, class Backend>void *Distributor<Answer,Qu
         return 0;
       }
       catch(...) {
-        L<<Logger::Error<<Logger::NTLog<<"Caught unknown exception in Distributor thread "<<(unsigned long)pthread_self()<<endl;
+        L<<Logger::Error<<Logger::NTLog<<"Caught unknown exception in Distributor thread "<<(unsigned int)pthread_self()<<endl;
 	delete b;
         return 0;
       }
